Create append function to add a node to the list
The append function adds a note to the end of
the list. First it checks whether the list is empty
-> if yeas then the head's next becomes the node.

Else it iterates over all the nodes untill it
finds the tail (the tail will have next === null).

Also added some checking for when our tmp is not
a object or if is null or if the tmp's next is
actually not null
- this should not even happen but still.